Move unexported symbols to lint_unexported.go Make attributeSet exported
