+++ lint/lint.go
@@ -1,31 +1,13 @@
 package lint
 
 import (
-	"fmt"
 	"go/ast"
 	"go/token"
 	"go/types"
-	"reflect"
 	"sort"
 
 	"github.com/go-toolsmith/astfmt"
 )
-
-type checkFunction interface {
-	New(*context) func(*ast.File)
-}
-
-type ruleInfo struct {
-	attrs attributeSet
-
-	New func(*context) func(*ast.File)
-}
-
-// checkFunctions is a table of all available check functions
-// as well as their metadata (like "experimental" attributes).
-//
-// Keys are rule names.
-var checkFunctions = map[string]*ruleInfo{}
 
 // RuleList returns a slice of all rules that can be used to create checkers.
 // Slice is sorted by rule names.
@@ -33,8 +15,8 @@ func RuleList() []*Rule {
 	rules := make([]*Rule, 0, len(checkFunctions))
 	for ruleName, info := range checkFunctions {
 		rules = append(rules, &Rule{
-			name:  ruleName,
-			attrs: info.attrs,
+			name:         ruleName,
+			AttributeSet: info.AttributeSet,
 		})
 	}
 	sort.SliceStable(rules, func(i, j int) bool {
@@ -43,11 +25,33 @@ func RuleList() []*Rule {
 	return rules
 }
 
+// AttributeSet describes rule implementation properties that may be
+// related to implementation (experimental) or be more fundamental (opinionated).
+type AttributeSet struct {
+	// SyntaxOnly marks rules that can be checker by using AST only.
+	// Such rule implementations should not use types info.
+	//
+	// It's OK not to mark rule as SyntaxOnly even if current
+	// implementation does not use types info.
+	SyntaxOnly bool
+
+	// Experimental marks rule implementation as experimental.
+	// Reasons to mark anything as experimental:
+	//	- rule name can change in near time
+	//	- checker gives more false positives than anticipated
+	//	- rule requires more testing or alternative design
+	Experimental bool
+
+	// VeryOpinionated marks rule as controversial for some audience and
+	// that it might be not suitable for everyone.
+	VeryOpinionated bool
+}
+
 // Rule describes a named check that can be performed by the linter.
 type Rule struct {
-	name string
+	AttributeSet
 
-	attrs attributeSet
+	name string
 }
 
 // String returns r short printed representation (name only).
@@ -55,21 +59,6 @@ func (r *Rule) String() string { return r.name }
 
 // Name returns rule name.
 func (r *Rule) Name() string { return r.name }
-
-// SyntaxOnly reports whether type info is not required to perform rule checks.
-func (r *Rule) SyntaxOnly() bool { return r.attrs.SyntaxOnly }
-
-// Experimental reports whether rule is experimental.
-//
-// Rules are considered experimental when they have many
-// false positives or some unresoleved bugs in them.
-func (r *Rule) Experimental() bool { return r.attrs.Experimental }
-
-// VeryOpinionated reports whether rule is very opinionated.
-//
-// Rules are considered opinionated when they might be controversial
-// and not suitable for broad audience.
-func (r *Rule) VeryOpinionated() bool { return r.attrs.VeryOpinionated }
 
 // NewChecker returns checker for the given rule.
 //
@@ -115,6 +104,7 @@ func (c *Checker) Check(f *ast.File) []Warning {
 // Warning represents issue that is found by rule checker.
 type Warning struct {
 	// Node is an AST node that caused warning to trigger.
+	// Can be used to obtain proper error location.
 	Node ast.Node
 
 	// Text is warning message without source location info.
@@ -137,43 +127,4 @@ type Context struct {
 	// SizesInfo carries alignment and type size information.
 	// Arch-dependent.
 	SizesInfo types.Sizes
-}
-
-// context is checker-local context copy.
-// Fields that are not from Context itself are writeable.
-type context struct {
-	*Context
-
-	// printer used to format warning text.
-	printer *astfmt.Printer
-
-	warnings []Warning
-}
-
-// Warn adds a Warning to checker output.
-func (ctx *context) Warn(node ast.Node, format string, args ...interface{}) {
-	ctx.warnings = append(ctx.warnings, Warning{
-		Text: ctx.printer.Sprintf(format, args...),
-		Node: node,
-	})
-}
-
-func addChecker(c checkFunction, attrs ...checkerAttribute) {
-	var info ruleInfo
-	for _, attr := range attrs {
-		switch attr {
-		case attrExperimental:
-			info.attrs.Experimental = true
-		case attrSyntaxOnly:
-			info.attrs.SyntaxOnly = true
-		case attrVeryOpinionated:
-			info.attrs.VeryOpinionated = true
-		default:
-			panic(fmt.Sprintf("unexpected checkerAttribute"))
-		}
-	}
-	typeName := reflect.ValueOf(c).Type().Name()
-	ruleName := typeName[:len(typeName)-len("Checker")]
-	info.New = c.New
-	checkFunctions[ruleName] = &info
 }
